GhostPDF - add control for Marked Content and document
Now that we can pass marked content from the input to the pdfwrite
device we need to be able to control it (this can significantly increase
the size of a PDF file).

This commit adds a new control to GhostPDF '-dPreserveMarkedContent'
which defaults to true. If false then marked content is not passed to
the pdfwrite device.

Add documentation for the parameter, fix formatting of the
PreserveAnnots and PreserveAnnotTypes while we're here.

Finally, when processing optional content, do not emit an EMC if it is
associated with optional content.