Only select cmap table for newly create Freetype fonts
The code selecting the cmap table was erroneously in the common between creating
a completely new Freetype font, and rescaling an existing one, which opened the
chance we'd end up with a cmap selected we did not expect.

Move that code into the block only for newly created Freetype fonts.