Bug 704680: Fix font loading info message file descriptor
The code writing font loading information to the back channel originally
only did so for debug builds, in (partially) converting it work on non-debug
builds, I'd accidentally had some of it writing to stderr - ending up with
some message parts going to stdout and some to stderr.

Fix that so it's consistent, and works with non-debug builds.