Fix for setting of PageSpotColors Bug 704660
plmain.c should not be setting the PageSpotColors
value.  This should be set by the individual
interpreters on a per page basis.  This change
fixes a crash (Bug 704660) reduces the number of
closing and opening times of the device with pdfi
and pdfi renders more than 4 spots to the separation
devices just fine. Also for gpdl with HEAD, the subsequent
execution of a PS file following a PCL file
did not properly reset the PageSpotColors to -1
as is required for PS files, resulting in improper
output when going to a separation device.  This commit
includes a fix for that.  Tested with command line

gpdl -sDEVICE=psdcmyk -r72  -o ./myoutputs/pdl_input_%d.psd
./myinputs/owl.pcl ./myinputs/Ad_InDesign.ps
./myinputs/spots_multi_page.pdf  ./myinputs/input.xps
./myinputs/deviceNImage.eps ./myinputs/DeviceN_20Colors.pdf
./myinputs/page_spots/Ad_InDesign.ps

to verify that the number of spots was getting properly
set with each interpreter and page (when applicable) change.