Support user specified Fontmap files in pdfi
Handle -sFONTMAP param in gpdl, and copy the relevant strings from the
Postscript world when pdfi is called from Ghostscript.

To make things a little simpler, there is also a tweak to the Postscript code,
so the FONTMAP parameter gets split into an array of strings during
initialisation, rather than at use time, which saves doing that process once
for Postscript and once for pdfi.