GhostPDF - clamp invalid constant alpha values
Bug 704700 "PDF to PNG - Dropping annotations"

The annotations all use an ExtGState which has CA and ca values of 255.
The spec is clear that valid values lie in the range 0.0 to 1.0.

While this is an example of an invalid file we can render it correctly
by clamping out of rang values to the top or bottom of the range, and
this commit does that, along with adding a new warning that this has
been done.

We don't plan to address this in the old PostScript-based interpreter.