Bug 704670: Patterns with Alphabits, Transparency, and Clist
When the pattern accumulator is active, it should not pass along
to the target device the special op gxdso_restrict_bbox.
The bbox values of the target
device have no relevance to the size of the pattern.  For this
bug, the special op call was occurring in
alpha_buffer_init and the bbox was getting
altered based upon the cropping values of the clist writer.
Based upon the values returned, the code was determining that the
pattern contents did not need to be drawn, which was incorrect.
There are two possible cases that can occur when dealing with
the pattern accumulator.  It can be either using a clist or
not using a clist.  If it is using a clist, the clist will
have an extent that is the same as a buffer would have had and
the special operation call from alpha_buffer_init will go
directly to the clist device and the clip adjusted correctly.
If instead the pattern accumulator is not using a clist, then
we are drawing the pattern into a buffer (with or without transparency).
In this case, the buffer is the target device for the path
we are drawing and gxdso_restrict_bbox should be handled by the
pattern accumulator.