GhostPDF - Fix psdcmyk device when page has too many spots
This is realted to the commit which sends SeparationColorNames to
the device using put_params().

The reason we pass SpotColorNames to the device is to cope with Zoltan's
file which sets a halftone with a spot colour, then draws content
in that spot colour.

Because we have not added the spot colour at the time the halftone is
set, we discard the component. Later when we add the spot colour due to
actually using it, we use the default halftone, not the correct one.

By setting the spot colour up front we avoid this problem.

However, perf-testing/pdf/J12_Acrobat.pdf has too many spots (it has 16)
and when we try to set that spot array we get an error. Ghostscript's
zputparams() operator simply ignores the errors (!!) while pdfi passes
them back, which causes it to essentially abort the page content.

But, if we send multiple parameters to the device, and one causes an
error, then it is possible that the remaining parameters are not
processed. So ignoring the error is not, by itself, sufficient.

This commit breaks out the SpotColorNames processing separately and
sends it separately too. If that fails we just ignore it. Otherwise
we properly handle the device being closed and reopen it in all cases.

This solves the problem with J12_Acrobat for me. I have a
suspicion that pdfi is still not properly handling psdcmyk with spot
files though.