Change how gs_offset_t is selected and add PRI macro for it
Originally, to change the gs_offset_t a port would specify GS_OFFSET_T as a
CFLAG, this changes it so the port would specify ARCH_SIZEOF_GS_OFFSET_T to be
either 4 or 8, which selects between int32_t and int64_t respectively.

That allows us to include a PRI style macro "PRIdOFFSET" for use as string
formatters, also added here.