GhostPDF - Use the PRIdOFFSET macro for gs_offset_t types
Technically the gs_offset_t can be anything, and in particular can be
either 32 or 64 bits. This makes it awkward to know what format
specifier to use for gs_sprintf and sscanf.

Since Chris added the new PRIdOFFSET macro we can now use that for both
cases. This makes it much less likely that we will try to read a 64-bit
value into a 32-bit variable.

While we're at it, change a few debug statements that were using the
(potentially) wrong format specifier to gs_sprintf.