GhostPDF - clamp path co-ordinates as per Ghostscript
Discovered this while chasing down a different problem. This commit
b535d8ca48b19eeff379706b4142c2f1a7216584 altered the way path
co-ordinates are handled, and does it *only* for Ghostscript.

It also only alters 'some' path construction. There is no information
in the commit about the 'clamp_coordinates' flag in the graphics state
either.

However, Ghostscript's initialisation (which differs from all the other
interpreters) sets the 'clamp_coordinates' flag in the graphics state
to true which means that some files which would otherwise throw an error
instead clamp the co-ordinate to the maximum/minimum.

Do the same here for GhostPDF. This alters the way a couple of files are
rendered at 300 dpi and higher, some content is now rendered which
previously caused a limitcheck error.