Pass x0,y0 to device color deserialisation (read) calls.
When we serialise colors into the clist, we want to include the
"phase" of colors with them where appropriate. During reading
the phase for such colors will need to be offset by the offset
of the screen.

If, for example, we are rendering for a band that starts at
(0, 1000) then the phase will need to be offset by 1000. Therefore
pass this offset to the color reading code, so we can adjust by
it.