GhostPDF - defer creation of dummy pattern cache entry for pdfwrite
This is exhibited by Bug693422.pdf, previously we were creating the
dummy cache entry for a high level pattern before we ran the pattern
PaintProc.

The problem occurs if, during the course of executing the Pattern
PaintProc, we run another pattern which happens to have an id which
'hashes' the same as the id for the outer pattern.

If that should happen then the inner Pattern would overwrite the slot
in the pattern cache being used by the outer pattern with the inner
Pattern's tile. This would then mean that we could not load the pattern
from the cache in order to use it.

So we would run the Pattern PaintProc again, presumably on the
assumption that the tile had been freed.

This would lead to an endless loop.