GhostPDF - add a page count to the output to match Ghostscript
This makes GhostPDF more closely match the Ghostscript output. I did
try to change the PL layer 'End of page...' message to the Ghostscript
'>>> howpage, press <return> to continue<<<' message but that turns
out not to be possible.

While we can define our own 'finish_page()' method we don't know whether
'-dNOPAUSE' was on the command line, because that is processed by the
PL layer before we pass parameters to the PDF interpreter. The PL
layer stores the value in a pl_main_instance_s structure, which is
defined in plmain.c. This means that although we can access the pointer
to the structure, we can't peek inside it.

So in order to have NOPUASE work, we need to call pl_main_finish_page()
and let that print the message.