pdfwrite - free XUID when invalidating
Fixes a memory leak in pdfi. Previously we used to just set the id
and xvalues of an XUID and relied on GC to clean up the allocation of
the xvalues array. That doesn't work if we aren't using GC'ed memory.

Copy the code in gsfont.c, gs_font_finalize() and in the case of an
XUID call uid_free() to free the xvalues.