Removal of globals in opvp device
Thanks to Robin for suggesting the structure form
to get the globals at a common offset for the
raster and vector forms of the device. That coupled
with pushing the device through all the methods and
doing initialization was the biggest effort.

Thanks also to Chris for his help in providing
understanding as to what this device was actually doing.

This commit includes a test harness for the opvp
device, which uses the device client API to write
out the commands it receives to a text file for
comparison.  To build the driver use  ./build_opv_harness.sh
To use the driver use a command line like
-sDEVICE=opvp -sDriver=libopv.so -o output.txt -f ./examples/tiger.eps
The command lines are always dumped to the same
file which is opvp_command_dump.txt.   Note that the
harness itself has to rely upon globals due to our
inability to change the client API.