GhostPDF - fix memory leak with colour spaces
If, while setting a colour space by name, we detect that the current
space is already the required space, then we short-circuit the setting
of the space.

But we were doing that after retrieving the object representing the
named colour space, and not counting down the object leading to a
memory leak.

Instead, we can check the space early and avoid creating the object
at all.