Fix ordering of separations when SeparationOrder not given
Commit c9e0e41129d90cc237c378c1042a2b063d6dc77f was supposed to do this
but it failed to account for the case where one spot name was a subset
of another spot name.

In this case we could end up with a broken ordering, with plates
missing or duplicated. This commit caters for spot names matching by
examining the lengths of the names as well.