Coverity IDs 374285, 374286 plus some tidying up of page reporting
Use the PRId64 macro to print out the page number, instead of %d

While we're here making changes; implement picking up -dQUIET in
pdfi. If QUIET is set, don't spam the page number to stdout. Remove the
emission of the page number from the PostScript code as pdfi is now
doing that, and having both leads to duplicate stdout output.