GhostPDF - guard against out of range /Size in xref streams
Bug 704724 "Buffer overflow in read_xref_stream_entries"

If the /Size entry in an xref stream was too large (when multiplied by
the size of an xref entry) to be represented as a size_t, then we could
'wrap round' the number and allocate too few bytes to hold the xref,
leading to a buffer overrun.

This is much more likely to happen with a 32-bit build but is
potentially with a 64-bit build as well.

The same is technically possible with a regular xref table too, so cater
for both cases here by checking the Size against the ARCH_MAX_SIZE_T
divided by the size of an xref entry. Use ceil() to round up.