Fix memory leaks with copied fonts
Copied fonts are used by pdfwrite (as the original font may have been
restored away when we output the file). For CIDFonts we were not freeing
the subrs of the (copied) descendant font when the copied font was
freed.

In addition for TrueType outline CIDFonts we were not freeing the
CIDSystemInfo strings (we did for other outline types).