Font copying - fix memory leak on CIDFont type 0
The comment in copy_font_cid0() says that the glyph data and global
Subrs of the descendant fonts are shared with the parent font, so we
overwrite the descendant font pointers.

For the glyph data and names we were freeing the copied data before we
overwrite it, for the global Subrs, we weren't leading to a memory
leak.