Fix some page mode vs clist halftone differences.
We were unpacking the screen phase from the clist and inverting it.
This works in many cases because the most common use of this is with
the code that writes it setting it to the device color phase, which
is in the opposite sense.

Unfortunately, not ALL places that wrote it were writing it in this way,
causing differences with PCL, particularly where rop texture devices
are in use.

Accordingly, we remove the inversion on read, and invert it at the
appropriate place when writing.

This produces ~85 cluster diffs (all bar one in the pbmraw device),
where the clist rendering now matches the page mode rendering.

The sole exceptions to this is tests_private/xl/pcl6cet3.0/c421.bin
to pbmraw at 600dpi where clist and pagemode differ on page 17 in
a slightly odd way.

This is so much of a progression that I have committed these fixes and
will look into that file afterwards.