Font copying - fix more memory leaks with CIDFonts
If we created a name for a CMap in a CIDFont, then we must free it or
leak memory. But if it's a standard name, we don't want to because
that comes from the interpreter's name table.

If we allocated Widths2 or 'v' (not sure what that is) arrays when we
copied a CIDFont, then we need to free them.