Bug 700556: Remove strange defs of .origdefinefont/.origundefinefont/.origfindfont
Nancy identified a strange set of definitions for:
/.origdefinefont
/.origundefinefont
/.origfindfont

with a comment saying it was to keep the original definitions in systemdict so
they would bind properly, which made no sense to any of us.

Process of elimination pointed to only /.origundefinefont being a problem, and
binding the definition /undefinefont in gs_dps1.ps made it no longer a problem.

This may need revisited if other problems arise. This change also appers to work
correctly with DELAYBIND.