pdfwrite - fix memory leak with manufactured glyph names
This appears to be a long standing memory leak in pdfwrite, which we
get away with in Ghostscript due to garbage collection.

The pdf_choose_output_glyph_name() procedure can, under certain
conditions, need to manufacture a glyph name because there isn't one
available. To do this it calls gs_alloc_string and returns the string
in the 'gs_constr_string' it was handed.

The problem is that nothing ever frees this memory.

Solving this is surprisingly difficult. I've removed the gs_const_string
from the pdf_encoding_element structure and replaced it with an explicit
data pointer and size, so that these can be allocated outside the
interpreter 'string' area (the existing code was causing warnings in a
debug build), and so that we can control the lifetime of the data.

I've also added a 'cleanup' flag to the function so that on return we
can tell that memory was allocated and stored in the string, and we can
call gs_free_string to discard it.

This does mean that instead of simply retaining a pointer to the string
we actually need to copy it. At some time in the past this was partially
done, but not for the encoding element. So add a function to copy the
data from a string into the members of the pdf_encoding_element,
allocating memory for it. Also add code to free the memory when we call
font_resoruce_free and free the Encoding.

Modify code in a couple of places that was relying on the member being
a gs_constr_string to instead use the member variables.

Finally, alter the allocation of the Encoding from a struct to a simple
byte pointer, and this allows me to remove the Garbage Collector
descriptors.

This seems to cure a number of memory leaks with pdfwrite and pdfi and
doesn't cause any cluster problems. Nor does it (on testing so far) add
any new memory leaks.