Fix one memory leak with 0693548_09210_Book4_edit_final.pdf
When setting the DefaultGray or DefaultCMYK color spaces we set the space's
interpreter_data to the PDF_NAME object of the color space and reference count
the name. Unfortunately, we weren't setting the callback function pointer to
deal with the interpreter_data when the color space was freed.

Setting that callback fixes that memory leaks.