pdfwrite - fix memory leak with non-existent pages
If a pdfmark has a /Dest with a page number beyond the page currently
being processed then we create a 'virtual' entry for it, which will
be updated later when we process the destination page.

However, if we never reach that page (bad pdfmark or the interpreter
was only processing a subset of pages) then we would fail to free the
virtual entry from the page list.