pdfwrite - fix memory leak with Encodings
It is possible (in an error case) to create an entry in an Encoding
for a simple font, but not to then use that entry. This leads to the
font FirstChar and LastChar potentially not including the Encoding
element leading to a leak on freeing the Encoding.

Fix by checking and freeing the entire possible range of Encoding
entries.