Update gx_clip_path to encapsulate fill_adjust too.
A gx_clip_path holds 2 possible representations of a clipping region.

Firstly, it will hold a list of rectangles that exactly cover the given
clipping region. Every gx_clip_path will hold such a list, though
in many cases it's a singleton.

Secondly, it may hold a gx_path pointer to point to the path that
can be used to regenerate the clipping list. In order for this
regeneration to happen correctly, we really need to know the
fill_adjust value that was in place when the path was added to the
clipping region.

This last piece of information was not kept by gx_clip_path - this
commit rectifies that.