pdfwrite - eliminate another gs_const_string case
The CMap name being stored for a type 0 composite font was running into
the same problem as other cases of gs_const_string. The code assumed it
would always get a string, but sometimes created one (using
gs_alloc_string), which would then leak without GC.

This commit just allocates memory and copies the string, freeing it when
the type 0 font is released. This cures a memory leak with pdfwtite
and pdfi.