GhostPDF - fix handling of JPX images
Previously we simply read data from the filtered stream until we had
read enough bytes to satisfy the image request, or ran out of data. For
most image formats this works acceptably, since we decode the compressed
data as we go.

For JPX images, however, this doesn't work. We try to read the entire
(compressed) image before we start to decode it. This means that for
JPX images we were reading from the start of the compressed image data
until we ran out of data to read. This wouldn't happen until we reached
the end of the PDF file.

For JPX passthrough this also meant we wrote all the data we had read as
well, to the output PDF file, resulting in potentially gigantic output
files. They still worked, because the JPXDecode filter doesn't care if
there is garbage after the content, but it was slow and producing
unreasonably large files.

The only solution to this is to use a SubFileDecode filter to limit the
amount of data we read. I've chosen to use an EOD string here, rather
than use the stream Length. That's because we know from previous
experience that Length keys can have incorrect values and can't be
trusted. If there's a missing endstream then we'll still stop as soon as
we either find another one or reach the end of file. At least we won't
drop anything.

This did, however, cause a different problem; the implementation of the
SubFileDcode was setting the filter 'min_size' to be the size of the
string. This doesn't work. The min_size variable actually appears to
limit the amount of data read by the filter at a time. In the case here
it limited a buffer of data to 9 bytes. If a higher filter in the chain
needed more data than that at once, then we would enter a loop where the
upper filter requested data, the SubFileDecode filter would return
because there was data still to be consumed and round and round we would
go.

Yet another example of the lunacy of the stream code.

Fixed by not setting min_size to the size of the string.