Bug 704775: Fix clist vs page mode rendering of shading.
Rendering tests_private/comparefiles/Bug691152.pdf at 300dpi to
pbmraw produces a different result in both page mode and clist
mode. A simplified file is given in the bug.

The problem is that clist rendering ends up decomposing the
shading under the 'pbmraw' device, which is not separable and
linear, rather than its parent 'pdf14clist_gray' device, which
is.

The fix is to lift a special case from gx_default_fill_path up to
pdf14_clist_fill_rectangle. In gx_default_fill_path, we spot
that a path is being filled by a shading or a pattern, and we
rework the problem from 'fill this path with this color' to be
'fill a rectangle with this color through a clip path formed from
that path'. This is much more efficient for the graphics library
to handle as it avoids us repeatedly having to decompose the
shading for each rectangle generated by the scanconversion.

This special case handling is done by a separate routine, called
gx_default_fill_path_shading_or_pattern. In pdf14_clist_fill_path
we spot that we have a type 2 pattern (the pattern type that
encapsulates shadings), and we call that same routine here.

This results in the decomposition happening with the pdf14clist_gray
device rather than any others.

This appears to cause several diffs in the cluster, but testing a
random selection shows these appear to pull clist and page mode into
alignment.