pdfwrite - fix memory leak with patterns and dashed lines
On exiting a pattern accumulation we reset the graphics state to the
gstate in force at the start of the pattern accumulation (any changes
to the graphics state during the course of the pattern belong to the
pattern).

If the pattern altered the line dash parameters without grestoreing them
away then we would end up with an allocated dash pattern which would
leak when we returned the graphics state to it's initial values.

So here, if the dash pattern is not NULL and differs from the initial
graphics state, then we free the object before we overwrite it.