GhostPDF - fix memory leak in self-referencing dictionaries
If a dictionary had a key whose value was an indirect reference, and the
reference was back to the dictionary, then we would end up with the
dictionary (quite correctly) having a reference count of at least 1
even when no other object was referencing it.

This causes a memory leak on exit. It doesn't seem reasonable that this
is ever correct, but this commit solves the self-reference by leaving
the key in the dictionary associated with an indirect object instead of
replacing it with the dereferenced object. We return the dereferenced
object so everything 'ought' to work normally.

We do signal an error for this.