OSS-fuzz #42282
A broken Object stream could be 'repaired' in such a way as to use
object numbers outside the current size of the xref table. This would
lead to corruption of memory because we wrote directly to the table
using the index. For me this caused a seg fault during repair.

Fix this by calling pdfi_repair_add_object() if the object number
exceeds the current size of the xref table, and then set the table
entry contents.