Stroke filled by pattern optimisation.
Ghostscript has long had an optimisation for paths filled with
patterns and shadings. Instead of scan converting the path, and
then 'filling' that with the pattern, the code would instead
make a clipping device from the path, and then fill that in one
hit using a rectangle fill.

The reason why this is a win can be seen easily when you consider
a shading fill; if we scan converted to get a series of rectangles
that we then filled, the shading would need to be decomposed
repeatedly for each rectangle. By scan converting, making a
clipping device from that, and filling a single rectangle of the
shading, the decomposition only happens once - and the lowest
level 'fills' resulting from that are efficiently filtered by
the clipping device.

The same optimisation has never been applied to stroking, despite
it potentially being even more worthwhile. This commit rectifies
this ommission.

Sadly, applying it causes content to vanish in clist mode. After
some digging it seems that the clist writing code has to wrap
path filling operations with some save/restore magic for cropping
min/max values. Once we extend this mechanism for stroked paths
we need to do the same for stroked operations too.