OSS-Fuzz - 42360 and 42342
If we get an unidentified colour space in a JPX decode, then treat that
as an error. If we fail to get a colour space for an image, and the
image is not a mask (which uses the current colour space) then treat
that as an error too.

Prevents us trying to dereference a NULL colour space later.

42342 - Is caused by Acrobat's insane tolerance for broken files.

If we have a Xobject with no Subtype we can't tell whether its a Form or
Image. This should be an error but Acrobat assumes its a Form and
carries on. If it's a Form it should be a stream, but we've got an
example of a file where the Form is a dictionary with a Contents key.
Which should be illegal but, of course, Acrobat opens it.

The problem here is that the Contents key points back to the enclosing
Page dictionary, and we end up in a loop.

Check the Parent (and all Parents) to see if they match the  Contents of
this 'Form'. If so then we will end up in a loop, so abort.