oss-fuzz 42294: Reset buffer length after rewriting xref entry
If we find an invalidly terminated xref entry, we try a slower method to derive
the correct values, and write those back to the buffer with the correct form.

But we were failing to reset the length (which was used as a counter) so the
next block of code could its job.