GhostPDF - OSS-Fuzz #492299
pdfi_read_token() was returning with a code of 0, but without pushing
an object on the stack (because we ran out of input). In many places
this is catered for, mostly fairly low-level instances, by checking
to see if the stack has any objects on it.

However it seems better to alter the way that pdfi_read_token() works
slightly. It now returns 1 if it read a token, < 0 for an error, and
0 if there was no error but it didn't read a token (ie ran out of
input).

All the places that call pdfi_read_token have been reviewed and the
return value checked. If necessary new action is taken on 0.