OSS-fuzz #42309
The stream interface needs us to handle single filters, and arrays of
filters differently in order to correctly clean up after an error.

However, Flate with Predictor, although a single filter, is actually
handled as 2 filters internally and so requires special casing. When we
get an error with the Predictor filter we were closing all the open
filters in the chain.

Previously this was acceptable but since adding a SubFileDecode around
images, this meant that an error on the Predictor would lead to us
closing all the filters, including the SubFileDecode filter, which
would then crash when we tried to close it in the image code.

Simply have the Predictor code only close the parent stream (the Flate)
and leave the rest up to the file code.