OSS-fuzz #42453
The fuzzed file ended up trying to deal with a broken token, and in the
process treated it as a sc instead of  an scn. This caused us to get the
number of components for a Pattern space, which is -1. That caused a
seg fault.

Fixed by checking the returned number of components and if it is -1
(a Pattern colour space) then throw an error for operators that don't
support Pattern spaces.