OSS-fuzz #42664 - Don't clear stack on failed array read
If we got a ']' we would try to construct an array from objects on the
stack. If there was no initial mark ('[') then that would fail, as it
should, but we would clear the stack and attempt a recovery.

This discarded objects expected by higher level code. So just return
the error, don't try to be clever, and let the callers cope with it.

May need to revisit this after more local file testing but the cluster
is happy with it.