OS-fuzz #42718 - don't use ctx->Trailer directly.
If we repair the file (which can happen any time) it is possible that
ctx->Trailer might be replaced. If we are using it, and have not taken
a separate reference, then it could vanish while we are using it.

Since this isn't the first time this has come up, review and amend all
places using ctx->Trailer to use a temporary reference instead. Add
comments in each location and add comments to the definition in
ghostpdf.h to try and avoid a repetition.