OSS-Fuzz #42720 - Don't try to dereference /Parent of a dictionary
This code looks to me like there are some likely further problems. It is
only (I think) used by the pdfmark code for passing 'metadata' to the
pdfwrite device, it appears to be trying to create a string
representation of complex PDF objects.

I don't thin this caters at all well for the possibility of circular
references in the objects, but I may be mistaken. In any event, one
thing we should not do is try to dereference the Parent of any
dictionary, because we are recursively dereferencing every object, so if
we follow this to the Parent object then we will inevitably end up back
at the child and will loop forever.

This commit fixes the problem exhibited by the OSS-fuzz file but I think
more work is needed here.