OSS-fuzz #42761 - Loosen the recursion test in composite objects
We already checked to see if an indirect object stored in a composite
object (array or dictionary) pointed to the enclosing object, but we
check the generation number as well as the object number.

That could lead to recursion still taking place if the generation
numbers differed. Since the object number is all that really matters
here, drop the check against the generation number.