GhostPDF - yet another filter chain fix
If we applied an image scaling filter we did not add it to the chain
of filter (if any) from the /Filter array specified in the image
dictionary.

This then mean that the cleanup code stopped at the wrong place when
following and closing the chain of filters. This commit updates the
chain by manually correcting the 'original' member of the newly created
stream to point to the top of the chain, effectively adding this fitler
to the chain.

This fixes a memory leak.

Note; we cannot do this with the SubFileDecode filter for non-inline
images, because we apply that before we create the filter chain. We
could update the chain after creation but that seems flaky and no better
than the current approach.