OSS-fuzz #42847 - more work on circular references with /Parent
After commit 2b8c6f669e9525fdc25aadffefdc5f8b978fb0e0 where I thought
the use of /Parent looked suspicious, this is a case where there is a
problem; there is a circular reference through the /Parent entry in
a page or stream dictionary.

I've fixed it here by adding the Parent object to the loop detection
(we normally only add objects while they are being dereferenced) which
meant that we needed to close the loop detection for XObjects before
processing the XObject (it was done afterwards previously to ease error
recovery).

I suspect this may still need some investigation and I'm considering
rewriting this part of the code a bit.