OSS-fuzz #43012 - validate transfer function before use.
In gs_begin_transparency_mask() the code uses any supplied transfer
function, and assumes it will be a 1-in, 1-out function, so it only
supplied a single float (rather than an array) for the output value.

But the code does not check the transfer function only has a single
output before calling the function to sample it, which would lead to the
evaluation code running off the end of the array.

Since the transparency code doesn't validate the function, do so in
the interpreter code. If the function isn't valid just ignore it.