OSS-fuzz #42957 - do not rely on stack reference
When creating a stream object to return, the code was using the
dictionary from the stack and relying on the reference held by the stack
to ensure that the dictionary didn't disappear.

But if the repair code is called it can clear the stack.

So instead keep the stream object counted up until we are done with it.