OSS-fuzz #43124 - detect circular references when ignoring free objects
We have a work around for attempting to dereference objects marked as
free in the xref, to deal with some badly broken Sumatra files.

However this OSS-fuzz file causes that code to end up in a circular
reference. We can solve that by adding the object number to the stack
of objects being checked, if we are using loop detection (which we
pretty much always should be here).