Fix a Coverity report - use of potentiallly uninitialised buffer
The pointer 'dp' is initialised in set_cmd_put_op, but we were not
checking the return value from that function until after we had copied
colour data into the returned pointer. If the function returned an
error code we would be trying to write to an uninitialised pointer.

Everywhere else checks the return value before attempting to write into
the buffer, so we do the same here.

Picked up and reported (along with the patch) by Richard Lescak of
RedHat.